Refactor and reorganize Bangladeshi language conversion files; remove deprecated files and add new mappings
