--- index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>BANGLADESH LANGUAGE is one kayboard layout use any devices</title>
-  </head>
-  <body>
-    <div
-      style="
-        display: flex;
-        align-items: center;
-        width: 100%;
-        height: 100vh;
-        color: black;
-      "
-    >
-      <h1 style="width: 100%; text-align: center; font-size: 5rem">
-        WELLCOME TO<br /><span
-          style="
-            background: linear-gradient(to right, red 0%, green 100%);
-            background-clip: text;
-            -webkit-text-fill-color: transparent;
-          "
-          >BANGLADESH LANGUAGE</span
-        >
-      </h1>
-    </div>
-  </body>
-</html>
