+++ src/common/bangladeshi-sign-alphabets.ts
