+++ src/common/bangladeshi-vowel-alphabets.ts
