+++ src/common/bangladeshi-vowel-sign-alphabets.ts
