+++ src/index.ts
@@ -1,6 +1,7 @@
 import { Hono } from 'hono';
-import BijoyToUnicode from './bijoy2unicode';
-import UnicodeToBijoy from './unicode2bijoy';
+import IndexPage from '../public/index';
+import BijoyToUnicode from './bijoy2unicode/bijoy2unicode';
+import UnicodeToBijoy from './unicode2bijoy/unicode2bijoy';
 
 type Bindings = {
 	CACHE: KVNamespace;
@@ -8,25 +9,8 @@ type Bindings = {
 
 const app = new Hono<{ Bindings: Bindings }>();
 
-app.get('/', async (c) => {
-	const username = 'toufikforyou';
-
-	const repsCashe = await c.env.CACHE.get(username, 'json');
-
-	if (repsCashe) {
-		return c.json(repsCashe);
-	}
-
-	const resp = await fetch(`https://api.github.com/users/${username}/repos`, {
-		headers: {
-			'User-Agent': 'Cf Worker',
-		},
-	});
-
-	const data = await resp.json();
-	await c.env.CACHE.put(username, JSON.stringify(data));
-
-	return c.json(resp);
+app.get('/', (c) => {
+	return c.html(IndexPage());
 });
 
 app.get('/bijoy2unicode/:text', (c) => {
